Crate libusb

This crate provides a safe wrapper around the native libusb library.

Structs

ConfigDescriptor
Describes a configuration.

Context
A libusb context.

Device
A reference to a USB device.

DeviceDescriptor
Describes a device.

DeviceHandle
A handle to an open USB device.

DeviceList
A list of detected USB devices.

Devices
Iterator over detected USB devices.

EndpointDescriptor
Describes an endpoint.

EndpointDescriptors
Iterator over an interface's endpoint descriptors.

Interface
A device interface.

InterfaceDescriptor
Describes an alternate setting for an interface.

InterfaceDescriptors
Iterator over an interface's descriptors.

Interfaces
Iterator over a configuration's interfaces.

Language
A language used to read string descriptors from USB devices.

LibraryVersion
A structure that describes the version of the underlying libusb library.

Version
A three-part version consisting of major, minor, and sub minor components.

Enums

Direction
Transfer and endpoint directions.

Error
Errors returned by the libusb library.

LogLevel
Library logging levels.

PrimaryLanguage
Primary language families.

Recipient
Recipients of control transfers.

RequestType
Types of control transfers.

Speed
Device speeds. Indicates the speed at which a device is operating.

SubLanguage
Language dialects and writing systems.

SyncType
Isochronous synchronization mode.

TransferType
An endpoint's transfer type.

UsageType
Isochronous usage type.

Functions

request_type
Builds a value for the bmRequestType field of a control transfer setup packet.

version
Returns a structure with the version of the running libusb library.

Type Definitions

Result
A result of a function that may return a Error.