Struct libusb::ConfigDescriptor

pub struct ConfigDescriptor {
    // some fields omitted
}

Describes a configuration.

Methods

impl ConfigDescriptor

fn number(&self) -> u8

Returns the configuration number.

fn max_power(&self) -> u16

Returns the device's maximum power consumption (in milliwatts) in this configuration.

fn self_powered(&self) -> bool

Indicates if the device is self-powered in this configuration.

fn remote_wakeup(&self) -> bool

Indicates if the device has remote wakeup capability in this configuration.

fn description_string_index(&self) -> Option<u8>

Returns the index of the string descriptor that describes the configuration.

fn num_interfaces(&self) -> u8

Returns the number of interfaces for this configuration.

fn interfaces(&self) -> Interfaces

Returns a collection of the configuration's interfaces.