Struct libusb::InterfaceDescriptor

pub struct InterfaceDescriptor<'a> {
    // some fields omitted
}

Describes an alternate setting for an interface.

Methods

impl<'a> InterfaceDescriptor<'a>

fn interface_number(&self) -> u8

Returns the interface's number.

fn setting_number(&self) -> u8

Returns the alternate setting number.

fn class_code(&self) -> u8

Returns the interface's class code.

fn sub_class_code(&self) -> u8

Returns the interface's sub class code.

fn protocol_code(&self) -> u8

Returns the interface's protocol code.

fn description_string_index(&self) -> Option<u8>

Returns the index of the string descriptor that describes the interface.

fn num_endpoints(&self) -> u8

Returns the number of endpoints belonging to this interface.

fn endpoint_descriptors(&self) -> EndpointDescriptors

Returns an iterator over the interface's endpoint descriptors.