Struct libusb::ConfigDescriptor

pub struct ConfigDescriptor { /* fields omitted */ }

Describes a configuration.

Methods

impl ConfigDescriptor

pub fn number(&self) -> u8

Returns the configuration number.

pub fn max_power(&self) -> u16

Returns the device's maximum power consumption (in milliwatts) in this configuration.

pub fn self_powered(&self) -> bool

Indicates if the device is self-powered in this configuration.

pub fn remote_wakeup(&self) -> bool

Indicates if the device has remote wakeup capability in this configuration.

pub fn description_string_index(&self) -> Option<u8>

Returns the index of the string descriptor that describes the configuration.

pub fn num_interfaces(&self) -> u8

Returns the number of interfaces for this configuration.

Important traits for Interfaces<'a>
impl<'a> Iterator for Interfaces<'a> type Item = Interface<'a>;

pub fn interfaces(&self) -> Interfaces

Returns a collection of the configuration's interfaces.