Struct libusb::Context

pub struct Context { /* fields omitted */ }

A libusb context.

Methods

impl Context

pub fn new() -> Result<Self>

Opens a new libusb context.

pub fn set_log_level(&mut self, level: LogLevel)

Sets the log level of a libusb context.

pub fn has_capability(&self) -> bool

pub fn has_hotplug(&self) -> bool

Tests whether the running libusb library supports hotplug.

pub fn has_hid_access(&self) -> bool

Tests whether the running libusb library has HID access.

pub fn supports_detach_kernel_driver(&self) -> bool

Tests whether the running libusb library supports detaching the kernel driver.

pub fn devices<'a>(&'a self) -> Result<DeviceList<'a>>

Returns a list of the current USB devices. The context must outlive the device list.

pub fn open_device_with_vid_pid<'a>(
&'a self,
vendor_id: u16,
product_id: u16
) -> Option<DeviceHandle<'a>>

Convenience function to open a device by its vendor ID and product ID.

This function is provided as a convenience for building prototypes without having to iterate a DeviceList. It is not meant for production applications.

Returns a device handle for the first device found matching vendor_id and product_id. On error, or if the device could not be found, it returns None.

pub fn register_callback(
&self,
vendor_id: Option<u16>,
product_id: Option<u16>,
class: Option<u8>,
callback: Box<Hotplug>
) -> Result<Registration>

pub fn unregister_callback(&self, reg: Registration)

pub fn handle_events(&self, timeout: Option<Duration>) -> Result<()>