Struct libusb::DeviceDescriptor

pub struct DeviceDescriptor {
    // some fields omitted
}

Describes a device.

Methods

impl DeviceDescriptor

fn usb_version(&self) -> Version

Returns the device's maximum supported USB version.

fn device_version(&self) -> Version

Returns the manufacturer's version of the device.

fn manufacturer_string_index(&self) -> Option<u8>

Returns the index of the string descriptor that contains the manufacturer name.

fn product_string_index(&self) -> Option<u8>

Returns the index of the string descriptor that contains the product name.

fn serial_number_string_index(&self) -> Option<u8>

Returns the index of the string descriptor that contains the device's serial number.

fn class_code(&self) -> u8

Returns the device's class code.

fn sub_class_code(&self) -> u8

Returns the device's sub class code.

fn protocol_code(&self) -> u8

Returns the device's protocol code.

fn vendor_id(&self) -> u16

Returns the device's vendor ID.

fn product_id(&self) -> u16

Returns the device's product ID.

fn max_packet_size(&self) -> u8

Returns the maximum packet size of the device's first endpoint.

fn num_configurations(&self) -> u8

Returns the number of config descriptors available for the device.