Struct libusb::DeviceDescriptor

pub struct DeviceDescriptor { /* fields omitted */ }

Describes a device.

Methods

impl DeviceDescriptor

pub fn usb_version(&self) -> Version

Returns the device's maximum supported USB version.

pub fn device_version(&self) -> Version

Returns the manufacturer's version of the device.

pub fn manufacturer_string_index(&self) -> Option<u8>

Returns the index of the string descriptor that contains the manufacturer name.

pub fn product_string_index(&self) -> Option<u8>

Returns the index of the string descriptor that contains the product name.

pub fn serial_number_string_index(&self) -> Option<u8>

Returns the index of the string descriptor that contains the device's serial number.

pub fn class_code(&self) -> u8

Returns the device's class code.

pub fn sub_class_code(&self) -> u8

Returns the device's sub class code.

pub fn protocol_code(&self) -> u8

Returns the device's protocol code.

pub fn vendor_id(&self) -> u16

Returns the device's vendor ID.

pub fn product_id(&self) -> u16

Returns the device's product ID.

pub fn max_packet_size(&self) -> u8

Returns the maximum packet size of the device's first endpoint.

pub fn num_configurations(&self) -> u8

Returns the number of config descriptors available for the device.