Struct libusb::EndpointDescriptor

pub struct EndpointDescriptor<'a> {
    // some fields omitted
}

Describes an endpoint.

Methods

impl<'a> EndpointDescriptor<'a>

fn address(&self) -> u8

Returns the endpoint's address.

fn number(&self) -> u8

Returns the endpoint number.

fn direction(&self) -> Direction

Returns the endpoint's direction.

fn transfer_type(&self) -> TransferType

Returns the endpoint's transfer type.

fn sync_type(&self) -> SyncType

Returns the endpoint's synchronisation mode.

The return value of this method is only valid for isochronous endpoints.

fn usage_type(&self) -> UsageType

Returns the endpoint's usage type.

The return value of this method is only valid for isochronous endpoints.

fn max_packet_size(&self) -> u16

Returns the endpoint's maximum packet size.

fn interval(&self) -> u8

Returns the endpoint's polling interval.