Struct libusb::EndpointDescriptor

pub struct EndpointDescriptor<'a> { /* fields omitted */ }

Describes an endpoint.

Methods

impl<'a> EndpointDescriptor<'a>

pub fn address(&self) -> u8

Returns the endpoint's address.

pub fn number(&self) -> u8

Returns the endpoint number.

pub fn direction(&self) -> Direction

Returns the endpoint's direction.

pub fn transfer_type(&self) -> TransferType

Returns the endpoint's transfer type.

pub fn sync_type(&self) -> SyncType

Returns the endpoint's synchronisation mode.

The return value of this method is only valid for isochronous endpoints.

pub fn usage_type(&self) -> UsageType

Returns the endpoint's usage type.

The return value of this method is only valid for isochronous endpoints.

pub fn max_packet_size(&self) -> u16

Returns the endpoint's maximum packet size.

pub fn interval(&self) -> u8

Returns the endpoint's polling interval.