Struct libusb::InterfaceDescriptor

pub struct InterfaceDescriptor<'a> { /* fields omitted */ }

Describes an alternate setting for an interface.

Methods

impl<'a> InterfaceDescriptor<'a>

pub fn interface_number(&self) -> u8

Returns the interface's number.

pub fn setting_number(&self) -> u8

Returns the alternate setting number.

pub fn class_code(&self) -> u8

Returns the interface's class code.

pub fn sub_class_code(&self) -> u8

Returns the interface's sub class code.

pub fn protocol_code(&self) -> u8

Returns the interface's protocol code.

pub fn description_string_index(&self) -> Option<u8>

Returns the index of the string descriptor that describes the interface.

pub fn num_endpoints(&self) -> u8

Returns the number of endpoints belonging to this interface.

Important traits for EndpointDescriptors<'a>
impl<'a> Iterator for EndpointDescriptors<'a> type Item = EndpointDescriptor<'a>;

pub fn endpoint_descriptors(&self) -> EndpointDescriptors

Returns an iterator over the interface's endpoint descriptors.